modified doctor post and move viewpost to postcolumn.kt
