+++ app/src/main/java/com/hellodoc/healthcaresystem/retrofit/RetrofitInstance.kt
@@ -19,7 +19,7 @@ import retrofit2.converter.gson.GsonConverterFactory
 object RetrofitInstance {
     //private const val BASE_URL = "http://192.168.0.104:4000"
 //    private const val BASE_URL = "http://192.168.100.3:3000"
-    private const val BASE_URL = "http://192.168.1.5:3000"
+    private const val BASE_URL = "http://192.168.1.7:4000"
 
     private val client = OkHttpClient.Builder()
         .build()
