+++ app/src/main/res/xml/network_security_config.xml
@@ -4,6 +4,6 @@
         <domain includeSubdomains="true">192.168.222.22</domain>
         <domain includeSubdomains="true">192.168.100.3</domain>
         <domain includeSubdomains="true">192.168.1.5</domain>
-<!--        <domain includeSubdomains="true">192.168.1.10</domain>-->
+        <domain includeSubdomains="true">192.168.1.7</domain>
     </domain-config>
 </network-security-config>
